Physics: Make a ship's position a vec<3, double>
! I hate to admit it, but a vec3 of singles is simply not exact enough;
  at least not unless one employs clever algorithms for coping with it
  (i.e. algorithms that are too clever for me).

+++ src/software.cpp
@@ -385,7 +385,7 @@ int ScenarioScript::luaw_set_ship_position(lua_State *ls)
                             sinf(lat),
                             cosf(lat) * cosf(lng),
                             1.f);
-    ship->position += height * ship->position.normalized();
+    ship->position += static_cast<double>(height) * ship->position.normalized();
 
     return 0;
 }
